refactor(contracts): promote canned phase prompts to @aep/contracts/prompts
Pure move of buildSpecGenerationInstruction/buildDesignGenerationInstruction
out of the console's promptStore into a shared source-exported module, so the
console CTAs and the root-level playground compose the same client half of
the instruction (docs/design/playground.md §9). No build step: prompt edits
apply on the next playground run. Also corrects the stale doc comment that
claimed console turns run useCase requirements-chat (they run as general).
