+++ packages/contracts/package.json
@@ -2,5 +2,16 @@
   "name": "@aep/contracts",
   "version": "0.0.0",
   "private": true,
-  "description": "Generated cross-service contracts. Holds the committed aep-api public OpenAPI spec (api/v1/openapi.yaml); regenerate via `make gen`."
+  "type": "module",
+  "description": "Generated cross-service contracts. Holds the committed aep-api public OpenAPI spec (api/v1/openapi.yaml); regenerate via `make gen`. Also the hand-written shared phase prompts (prompts/).",
+  "exports": {
+    "./prompts": "./prompts/index.ts"
+  },
+  "scripts": {
+    "typecheck": "tsc -p tsconfig.json --noEmit",
+    "lint": "eslint ."
+  },
+  "devDependencies": {
+    "typescript": "^5.9.3"
+  }
 }
